backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pandas==2.2.2
openpyxl==3.1.5
pdfplumber==0.11.4
odfpy==1.4.1
anthropic==0.34.2
python-dotenv==1.0.1
xlrd==2.0.1
sentence-transformers
